+++ src/lumicks/pyoptics/farfield_transform/__init__.py
@@ -6,68 +6,72 @@ References:
 This implementation is original code and not based on any other software
 """
 
-import numpy as np
+from typing import Tuple
 
-from ..mathutils.czt import czt
+import numpy as np
+from numpy.typing import ArrayLike
+from scipy.signal import CZT
 
 
 def czt_nf_to_ff(
-    Ex: np.ndarray,
-    Ey: np.ndarray,
-    Ez: np.ndarray,
-    sampling_distance: float,
+    Ex: ArrayLike,
+    Ey: ArrayLike,
+    Ez: ArrayLike,
+    sampling_period: float,
+    radius: float,
+    acceptance_angle: float,
     lambda_vac: float,
     n_medium: float,
-    n_bfp: float,
-    focal_length: float,
-    NA: float,
-    bfp_sampling_n=101,
-):
-    assert NA <= n_medium, "NA cannot be larger than n_medium"
-    Ex = np.atleast_2d(Ex)
-    Ey = np.atleast_2d(Ey)
-    Ez = np.atleast_2d(Ez)
-    assert Ex.shape == Ey.shape == Ez.shape, "All fields need to be equal size"
-    assert Ex.shape[0] == Ex.shape[1], "Field matrices need to be square"
-    dx = sampling_distance
+    sampling_distance: float = 0.0,
+    farfield_sampling_n=101,
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+    Ex, Ey, Ez = [np.atleast_2d(E) for E in (Ex, Ey, Ez)]
 
-    f = focal_length
+    if not 0 <= acceptance_angle <= np.pi / 2:
+        raise ValueError("The value for acceptance_angle needs to be between 0 and π/2")
+
+    if not (Ex.shape == Ey.shape == Ez.shape):
+        raise RuntimeError("All fields need to be equal size")
+
+    if not (Ex.shape[0] == Ex.shape[1]):
+        raise RuntimeError("Field matrices need to be square")
+    dx = sampling_period
+    sin_acceptance_angle = np.sin(acceptance_angle)
+
     k0 = 2 * np.pi / lambda_vac
     k = k0 * n_medium
-    ks = 2 * np.pi / sampling_distance
-    a = np.exp(-2j * np.pi * k / ks * NA / n_medium)
-    w = np.exp(-4j * np.pi * k / ks * (NA / n_medium) / (bfp_sampling_n - 1))
+    ks = 2 * np.pi / sampling_period
+    a = np.exp(-2j * np.pi * k / ks * sin_acceptance_angle)
+    w = np.exp(-4j * np.pi * k / ks * sin_acceptance_angle / (farfield_sampling_n - 1))
 
-    # The chirp z transform assumes data starting at x[0], but we don't know
-    # where point (0,0) is. Therefore, fix the phases after the
-    # transform such that the real and imaginary parts of the fields are what
-    # they need to be
+    # The chirp z transform assumes data starting at x[0], but that's not where the origin is.
+    # Therefore, fix the phases after the transform such that the real and imaginary parts of the
+    # fields are what they need to be
     samples_correction = (Ex.shape[0] - 1) / 2
-    phase_fix = (a * w ** -(np.arange(bfp_sampling_n))) ** samples_correction
-    phase_fix = phase_fix.reshape((bfp_sampling_n, 1))
+    phase_fix = (a * w ** -(np.arange(farfield_sampling_n))) ** samples_correction
+    x_czt = CZT(Ex.shape[0], farfield_sampling_n, w, a)
+    y_czt = CZT(Ex.shape[1], farfield_sampling_n, w, a)
 
-    phase_fix_1 = np.tile(phase_fix, (1, Ex.shape[0]))
-    phase_fix_2 = np.tile(phase_fix, (1, bfp_sampling_n))
+    fEx = x_czt(Ex, axis=0) * phase_fix[:, np.newaxis]
+    fEx = y_czt(fEx, axis=1) * phase_fix[np.newaxis, :]
+    fEy = x_czt(Ey, axis=0) * phase_fix[:, np.newaxis]
+    fEy = y_czt(fEy, axis=1) * phase_fix[np.newaxis, :]
+    fEz = x_czt(Ez, axis=0) * phase_fix[:, np.newaxis]
+    fEz = y_czt(fEz, axis=1) * phase_fix[np.newaxis, :]
 
-    fEx = np.transpose((czt(Ex, bfp_sampling_n, w, a)) * phase_fix_1, (1, 0))
-    fEx = czt(fEx, bfp_sampling_n, w, a) * phase_fix_2
-    fEy = np.transpose((czt(Ey, bfp_sampling_n, w, a)) * phase_fix_1, (1, 0))
-    fEy = czt(fEy, bfp_sampling_n, w, a) * phase_fix_2
-    fEz = np.transpose((czt(Ez, bfp_sampling_n, w, a)) * phase_fix_1, (1, 0))
-    fEz = czt(fEz, bfp_sampling_n, w, a) * phase_fix_2
-
-    sp = np.linspace(-NA / n_medium, NA / n_medium, bfp_sampling_n)
+    sp = np.linspace(-sin_acceptance_angle, sin_acceptance_angle, farfield_sampling_n)
     Sx, Sy = np.meshgrid(sp, sp)
     Sp = np.hypot(Sx, Sy)
     Sz = np.zeros(Sp.shape)
-    Sz[Sp <= NA / n_medium] = (1 - Sp[Sp <= NA / n_medium] ** 2) ** 0.5
+    Sz = (1 + 0j - Sp**2) ** 0.5
 
-    factor = -(2j * np.pi * k * Sz * np.exp(1j * k * f) * dx**2 / (f * 4 * np.pi**2))
-    Ex_inf = factor * fEx
-    Ey_inf = factor * fEy
-    Ez_inf = factor * fEz
+    factor = -(
+        2j * np.pi * k * Sz * np.exp(1j * k * radius) * dx**2 / (radius * 4 * np.pi**2)
+    ) * np.exp(-1j * k * Sz * sampling_distance)
+    factor[Sz.imag != 0.0] = 0.0
+    Ex_inf, Ey_inf, Ez_inf = [factor * fE for fE in (fEx, fEy, fEz)]
 
-    return ff_to_bfp(Ex_inf, Ey_inf, Ez_inf, Sx, Sy, Sz, n_medium, n_bfp)
+    return Sx, Sy, Sz, Ex_inf, Ey_inf, Ez_inf
 
 
 def ff_to_bfp(